+++ datanode/utils/observer.go
@@ -24,117 +24,172 @@ import (
 )
 
 type subscriber[T any] struct {
-	ch chan []T
+	ch chan T
 }
 
-type Observer[T any] struct {
+// Notifier is a simple pub/pub construct. Create one, and call Notify(...) to push values into it.
+// Subscribers register their interest by calling Subscribe() and are returned a channel which
+// will receive all the values that are sent by Notify() calls.
+//
+// Importantly, the notification is non-blocking. You can specify the buffer size for
+// the channels created inSubscribe() with 'chSize'.
+// If the channel is full when a notification is received the channel will be closed and the
+// subscription automatically unregistered.
+type Notifier[T any] struct {
 	subCount    int32
 	lastSubID   uint64
 	name        string
 	log         *logging.Logger
 	subscribers map[uint64]subscriber[T]
-	mut         sync.RWMutex
-	inChSize    int
-	outChSize   int
+	mu          sync.Mutex
+	chSize      int
 }
 
-func NewObserver[T any](name string, log *logging.Logger, inChSize, outChSize int) Observer[T] {
-	return Observer[T]{
+func NewNotifier[T any](name string, log *logging.Logger, chSize int) Notifier[T] {
+	return Notifier[T]{
 		name:        name,
 		log:         log,
 		subscribers: map[uint64]subscriber[T]{},
-		inChSize:    inChSize,
-		outChSize:   outChSize,
+		chSize:      chSize,
 	}
 }
 
-func (o *Observer[T]) Subscribe(ctx context.Context, filter func(T) bool) (chan []T, uint64) {
-	o.mut.Lock()
-	defer o.mut.Unlock()
+func (o *Notifier[T]) Subscribe() (chan T, uint64) {
+	o.mu.Lock()
+	defer o.mu.Unlock()
 
-	ch := make(chan []T, o.inChSize)
+	ch := make(chan T, o.chSize)
+	return ch, o.register(ch)
+}
+
+// Subscribe and immediately push 'msg' onto the channel as the first notification; handy for
+// sending a 'initial value' message to subscribers. The push onto the channel is non-blocking
+// and if 'inChSize' is not at least 1, the subscription will fail and an error will be returned.
+func (o *Notifier[T]) SubscribeAndNotify(msg T) (chan T, uint64, error) {
+	o.mu.Lock()
+	defer o.mu.Unlock()
+
+	ch := make(chan T, o.chSize)
+	select {
+	case ch <- msg:
+	default:
+		close(ch)
+		return ch, 0, fmt.Errorf("unable to push initial notification")
+	}
+
+	return ch, o.register(ch), nil
+}
+
+func (o *Notifier[T]) register(ch chan T) uint64 {
 	o.lastSubID++
 	o.subscribers[o.lastSubID] = subscriber[T]{ch}
-
-	ip, _ := contextutil.RemoteIPAddrFromContext(ctx)
-	o.logDebug(ip, o.lastSubID, "new subscription")
-	return ch, o.lastSubID
+	atomic.AddInt32(&o.subCount, 1)
+	return o.lastSubID
 }
 
-func (o *Observer[T]) Unsubscribe(ctx context.Context, ref uint64) error {
-	o.mut.Lock()
-	defer o.mut.Unlock()
-
-	ip, _ := contextutil.RemoteIPAddrFromContext(ctx)
-
-	if len(o.subscribers) == 0 {
-		o.logDebug(ip, ref, "un-subscribe called but, no subscribers connected")
-		return nil
-	}
+func (o *Notifier[T]) Unsubscribe(ref uint64) error {
+	o.mu.Lock()
+	defer o.mu.Unlock()
 
 	if sub, exists := o.subscribers[ref]; exists {
-		close(sub.ch)
-		delete(o.subscribers, ref)
+		o.unsubscribe(ref, sub)
 		return nil
 	}
 
 	return fmt.Errorf("no subscriber with id: %d", ref)
 }
 
-func (o *Observer[T]) GetSubscribersCount() int32 {
+func (o *Notifier[T]) UnsubscribeAll() {
+	o.mu.Lock()
+	defer o.mu.Unlock()
+
+	for ref, sub := range o.subscribers {
+		o.unsubscribe(ref, sub)
+	}
+}
+
+func (o *Notifier[T]) unsubscribe(ref uint64, sub subscriber[T]) {
+	// mutex must already be obtained before calling this function
+	close(sub.ch)
+	delete(o.subscribers, ref)
+	atomic.AddInt32(&o.subCount, -1)
+}
+
+func (o *Notifier[T]) GetSubscribersCount() int32 {
+	// Use atomic for GetSubscribersCount so we don't have to obtain the mutex to report it.
 	return atomic.LoadInt32(&o.subCount)
 }
 
-func (o *Observer[T]) Notify(values []T) {
-	o.mut.Lock()
-	defer o.mut.Unlock()
+func (o *Notifier[T]) Notify(value T) {
+	o.mu.Lock()
+	defer o.mu.Unlock()
 
 	if len(o.subscribers) == 0 {
 		return
 	}
 
-	if len(values) == 0 {
-		return
-	}
-
 	for id, sub := range o.subscribers {
 		select {
-		case sub.ch <- values:
-			o.logDebug("", id, "channel updated successfully")
+		case sub.ch <- value:
 		default:
-			o.logWarning("", id, "channel could not be updated, closing")
-			delete(o.subscribers, id) // safe to delete from map while iterating
-			close(sub.ch)
+			o.log.Warn(
+				fmt.Sprintf("%s notifier: %s", o.name, "channel could not be updated, closing"))
+			o.unsubscribe(id, sub)
 		}
 	}
 }
 
+// Observer build on Notifier - adding a method called Observe() which starts a goroutine that
+// - subscribes to the notifier
+// - optionally filters values, outputs them on a separate channel.
+// - retries a specified number of tries to output on the output channel before giving up.
+//
+// An Observer[T] uses a Notifier[]T to allow updates to be pushed through in batches.
+type Observer[T any] struct {
+	notifier  Notifier[[]T]
+	name      string
+	log       *logging.Logger
+	outChSize int
+}
+
+func NewObserver[T any](name string, log *logging.Logger, inChSize, outChSize int) Observer[T] {
+	return Observer[T]{
+		notifier:  NewNotifier[[]T](name, log, inChSize),
+		name:      name,
+		log:       log,
+		outChSize: outChSize,
+	}
+}
+
+func (o *Observer[T]) Notify(values []T) {
+	o.notifier.Notify(values)
+}
+
+func (o *Observer[T]) GetSubscribersCount() int32 {
+	return o.notifier.GetSubscribersCount()
+}
+
 func (o *Observer[T]) Observe(ctx context.Context, retries int, filter func(T) bool) (<-chan []T, uint64) {
 	out := make(chan []T, o.outChSize)
-	in, ref := o.Subscribe(ctx, filter)
+	in, ref := o.notifier.Subscribe()
 	ip, _ := contextutil.RemoteIPAddrFromContext(ctx)
 
 	go func() {
-		atomic.AddInt32(&o.subCount, 1)
-		defer atomic.AddInt32(&o.subCount, -1)
-
-		ctx, cancel := context.WithCancel(ctx)
-		defer cancel()
+		defer close(out)
 
 		for {
 			select {
 			case <-ctx.Done():
 				o.logDebug(ip, ref, "closed connection")
-				if err := o.Unsubscribe(ctx, ref); err != nil {
+				if err := o.notifier.Unsubscribe(ref); err != nil {
 					o.logError(ip, ref, "failure un-subscribing when context.Done()")
 				}
-				close(out)
 				return
 
 			case values, ok := <-in:
 				if !ok {
-					// 'in' channel may get closed because Notify() couldn't write to it
-					close(out)
+					// if 'in' channel is closed, it's because Notify() couldn't write to it.
+					// In that case we have already unsubscribed, so don't try again.
 					return
 				}
 				filtered := make([]T, 0, len(values))
@@ -153,7 +208,6 @@ func (o *Observer[T]) Observe(ctx context.Context, retries int, filter func(T) b
 					case out <- filtered:
 						retryCount = retries
 						success = true
-						o.logDebug(ip, ref, "sent successfully")
 					default:
 						retryCount--
 						if retryCount > 0 {
@@ -164,7 +218,10 @@ func (o *Observer[T]) Observe(ctx context.Context, retries int, filter func(T) b
 				}
 				if !success && retryCount <= 0 {
 					o.logWarning(ip, ref, "hit the retry limit")
-					cancel()
+					if err := o.notifier.Unsubscribe(ref); err != nil {
+						o.logError(ip, ref, "failure un-subscribing after send retry limit")
+					}
+					return
 				}
 			}
 		}
@@ -182,7 +239,7 @@ func (o *Observer[T]) logDebug(ip string, ref uint64, msg string) {
 }
 
 func (o *Observer[T]) logWarning(ip string, ref uint64, msg string) {
-	o.log.Debug(
+	o.log.Warn(
 		fmt.Sprintf("%s subscriber: %s", o.name, msg),
 		logging.Uint64("id", ref),
 		logging.String("ip-address", ip),